Move states to another package
