+++ src/main/java/akademia/ox/states/GameState.java
@@ -1,4 +1,4 @@
-package akademia.ox;
+package akademia.ox.states;
 
 public interface GameState {
     GameState moveToNextState();
